requirements.txt
PyMuPDF
camelot-py[cv]
pandas
weaviate-client
opencv-python
sentence-transformers
unstructured[pdf]
pdfplumber
openai
pillow
transformers
torch
torchvision
python-dotenv
langchain
langchain-community
langchain-weaviate